modularize code
separate LLM and database code
